# Backend/java/src/main/resources/static/swagger.yaml
openapi: "3.0.3"
info:
  title: "NoteS API"
  version: "v 0.0.1"
servers:
  - url: "{scheme}://{domain}:{port}/{context}"
    description: Resource server
    variables:
      scheme:
        enum:
          - 'http'
          - 'https'
        default: 'http'
      domain:
        enum:
          - 'localhost'
          - 'notes.fvds.ru'
        default: 'localhost'
      port:
        enum:
          - '8080'
          - '9080'
          - '9180'
          - '5267'
          - '8443'
          - '80'
          - '443'
        default: '5267'
      context:
        enum:
          - ''
          - 'server'
        default: ''
components:
  schemas:
    NoteSearchTagsResponseDto:
      required:
        - created_at
        - description
        - is_public
        - note_type
        - path
        - syntax_name
        - title
      type: object
      properties:
        description:
          type: string
        title:
          type: string
        path:
          type: string
        syntax_name:
          type: string
          enum:
            - PLAINTEXT
            - MARKDOWN
        note_type:
          type: string
          enum:
            - NOTE
            - COMMENT
            - COMMENT_REDACTED
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
    PaginatedNoteSearchResponse:
      type: object
      required:
        - items
        - total_pages
        - total_elements
        - page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NoteSearchTagsResponseDto'
        total_pages:
          type: integer
        total_elements:
          type: integer
          format: int64
        page:
          type: integer
...
    NoteCreateRequestDto:
      required:
        - syntax_name
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        content:
          type: string
          nullable: true
      additionalProperties: false
    NoteCreateResponseDto:
      required:
        - note_type
        - syntax_name
      type: object
      properties:
        path:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        note_type:
          "$ref": "#/components/schemas/NoteTypeName"
        is_public:
          type: boolean
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
      additionalProperties: false
    NoteEditRequestDto:
      type: object
      properties:
        content:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      additionalProperties: false
    NoteEditResponseDto:
      required:
        - syntax_name
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
      additionalProperties: false
    NoteEditPublicRequestDto:
      type: object
      properties:
        is_public:
          type: boolean
      additionalProperties: false
    NoteEditPublicResponseDto:
      type: object
      properties:
        title:
          type: string
          nullable: true
        is_public:
          type: boolean
      additionalProperties: false
...
    NoteTypeName:
      enum:
        - NOTE
        - COMMENT
        - COMMENT_REDACTED
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    SyntaxTypeName:
      enum:
        - PLAINTEXT
        - MARKDOWN
      type: string
...
  securitySchemes:
    Bearer:
      type: oauth2
      description: API использует [OAuth2 authorizationCode](https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/token
          scopes:
            read: Balea Server HTTP Api
security:
  - Bearer: [ ]
paths:
  "/api/private/{accountName}/notes/{pathNote}/publish":
    patch:
      tags:
        - PrivateNote
      description: Редактирование свойства публичности любой заметки или комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/NoteEditPublicResponseDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/NoteEditPublicResponseDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/NoteEditPublicResponseDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/private/{accountName}/notes/{pathNote}":
    patch:
      tags:
        - PrivateNote
      description: Редактирование любой заметки или комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/NoteEditResponseDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/NoteEditResponseDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/NoteEditResponseDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    get:
      tags:
        - PrivateNote
      description: Получение данных любой заметки или комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/NoteCreateResponseDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/NoteCreateResponseDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/NoteCreateResponseDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...
  "/api/public/{accountName}/notes/{pathNote}/publish":
    post:
      tags:
        - PublicNote
      description: Редактирование публичности заметки/комменатрия
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteEditPublicRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes/{pathNote}":
    post:
      tags:
        - PublicNote
      description: Редактирование заметки, комментарий в другом месте.
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - PublicNote
      description: Удаление заметки
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    get:
      tags:
        - PublicNote
      description: Данные заметки или комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes":
    post:
      tags:
        - PublicNote
      description: Создание заметки
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteCreateRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteCreateRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteCreateRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    get:
      tags:
        - PublicNote
      description: Список заметок пользователя
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 0
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            maximum: 20
            minimum: 5
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes/search/title":
    get:
      tags:
        - PublicNote
      description: Поиск по заголовку
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: Title
          in: query
          schema:
            type: string
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 0
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            maximum: 20
            minimum: 5
            type: integer
            format: int32
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNoteSearchResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes/search/tag":
    get:
      tags:
        - PublicNote
      description: Поиск по тегам
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: tag
          in: query
          schema:
            type: array
            items:
              type: string
        - name: filter
          in: query
          schema:
            type: array
            items:
              type: string
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 0
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            maximum: 20
            minimum: 5
            type: integer
            format: int32
        - name: and
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes/search/semantic":
    get:
      tags:
        - PublicNote
      description: Семантический поиск
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: Query
          in: query
          schema:
            type: string
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 0
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            maximum: 20
            minimum: 5
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...